# feat(memory, core, query): ChunkedWindowIterator for up to 7.5x faster queries