--- a/core/src/main/scala/filodb.core/query/PartitionTimeRangeReader.scala
+++ b/core/src/main/scala/filodb.core/query/PartitionTimeRangeReader.scala
@@ -69,11 +69,7 @@ final class PartitionTimeRangeReader(part: ReadablePartition,
     endRowNo = if (endTime >= info.endTime) {
                  info.numRows - 1
                } else {
-                 val result = timeReader.binarySearch(timeVector, endTime)
-                 // no match - binarySearch returns the row # _after_ the searched key.
-                 // So if key is less than the first item then 0 is returned since 0 is after the key.
-                 // Since this is the ending row inclusive we need to decrease row # - cannot include next row
-                 if (result < 0) (result & 0x7fffffff) - 1 else result
+                 timeReader.ceilingIndex(timeVector, endTime)
                }
   }