--- a/core/src/main/scala/filodb.core/store/ChunkSetInfo.scala
+++ b/core/src/main/scala/filodb.core/store/ChunkSetInfo.scala
@@ -117,6 +117,7 @@ object ChunkSetInfo extends StrictLogging {
 
   val emptySkips = new SkipMap()
   val log = logger
+  val nullInfo = ChunkSetInfo(UnsafeUtils.ZeroPointer.asInstanceOf[NativePointer])
 
   /**
    * ChunkSetInfo metadata schema: