--- a/jmh/src/main/scala/filodb.jmh/BasicFiloBenchmark.scala
+++ b/jmh/src/main/scala/filodb.jmh/BasicFiloBenchmark.scala
@@ -7,7 +7,7 @@ import scala.language.postfixOps
 import org.openjdk.jmh.annotations._
 import scalaxy.loops._
 
-import filodb.memory.format.vectors.LongBinaryVector
+import filodb.memory.format.vectors._
 import filodb.memory.NativeMemoryManager
 
 /**
@@ -32,6 +32,14 @@ class BasicFiloBenchmark {
   randomLongs.foreach(ivbuilder.addData)
   val iv = ivbuilder.optimize(memFactory)
   val ivReader = LongBinaryVector(iv)
+
+  val dblBuilder = DoubleVector.appendingVectorNoNA(memFactory, numValues)
+  randomLongs.map(_.toDouble).foreach(dblBuilder.addData)
+  val dblReader= dblBuilder.reader.asDoubleReader
+
+  val intBuilder = IntBinaryVector.appendingVectorNoNA(memFactory, numValues)
+  randomLongs.map(n => (n / 256).toInt).foreach(intBuilder.addData)
+  val intReader = intBuilder.reader.asIntReader
 
   val byteIVBuilder = LongBinaryVector.appendingVectorNoNA(memFactory, numValues)
   randomLongs.zipWithIndex.map { case (rl, i) => i * 10000 + (rl % 128) }.foreach(byteIVBuilder.addData)
@@ -74,6 +82,20 @@ class BasicFiloBenchmark {
   @OutputTimeUnit(TimeUnit.MICROSECONDS)
   def sumAllLongsSumMethod(): Double = {
     ivReader.sum(iv, 0, numValues - 1)
+  }
+
+  @Benchmark
+  @BenchmarkMode(Array(Mode.AverageTime))
+  @OutputTimeUnit(TimeUnit.MICROSECONDS)
+  def sumDoublesSumMethod(): Double = {
+    dblReader.sum(dblBuilder.addr, 0, numValues - 1)
+  }
+
+  @Benchmark
+  @BenchmarkMode(Array(Mode.AverageTime))
+  @OutputTimeUnit(TimeUnit.MICROSECONDS)
+  def sumAllIntsSumMethod(): Long = {
+    intReader.sum(intBuilder.addr, 0, numValues - 1)
   }
 
   @Benchmark