… README.md
… coordinator/src/test/scala/filodb.coordinator/NodeCoordinatorActorSpec.scala
… core/src/test/scala/filodb.core/memstore/TimeSeriesPartitionSpec.scala
… doc/query-engine.md
… memory/src/main/scala/filodb.memory/data/OffheapLFSortedIDMap.scala
… memory/src/main/scala/filodb.memory/format/vectors/DeltaDeltaVector.scala
… memory/src/main/scala/filodb.memory/format/vectors/DoubleVector.scala
… memory/src/main/scala/filodb.memory/format/vectors/IntBinaryVector.scala
… memory/src/main/scala/filodb.memory/format/vectors/LongBinaryVector.scala
… memory/src/test/scala/filodb.memory/format/vectors/DoubleVectorTest.scala
… memory/src/test/scala/filodb.memory/format/vectors/LongVectorTest.scala
… query/src/main/scala/filodb/query/exec/PeriodicSamplesMapper.scala
… query/src/main/scala/filodb/query/exec/rangefn/AggrOverTimeFunctions.scala
… query/src/main/scala/filodb/query/exec/rangefn/RangeFunction.scala
… query/src/test/scala/filodb/query/exec/LastSampleFunctionSpec.scala
… query/src/test/scala/filodb/query/exec/SlidingWindowIteratorSpec.scala
… query/src/test/scala/filodb/query/exec/WindowIteratorSpec.scala
… query/src/test/scala/filodb/query/exec/rangefn/AggrOverTimeFunctionsSpec.scala