**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [ ] Docs have been added / updated (for bug fixes / features) ?

**Current behavior :** (link exiting issues here : https://help.github.com/articles/basic-writing-and-formatting-syntax/#referencing-issues-and-pull-requests)

Currently almost all queries (except raw ones) use the `SlidingWindowIterator` for windowing operations.  It goes through every "row" in each Time series, spanning the first row of the earliest time window to the last row of the last time window, regardless of the step size or window length.  This works but does not take advantage of various optimizations that are possible with columnar layout.

**New behavior :**

This PR introduces a couple new classes/concepts:
- A `ChunkedWindowIterator` which gives the parameters of each window plus each applicable chunkSet within that window, instead of a sliding window row-based abstraction, directly to a 
- `ChunkedRangeFunction` which calculates and updates its results one chunk * window at a time

### Improvements in Query Speed

The following functions have `ChunkedRangeFunctions` implemented:
- sum_over_time
- count_over_time - optimized to just use binary search on timestamp vectors, no need to look at the value vector at all
- avg_over_time
- stdvar_over_time
- dev_over_time
- LastSampleFunction
- min_over_time
- max_over_time

Using the `QueryInMemoryBenchmark` specifically on `sum_over_time`, we can see the following improvements depending on window size and step, varying from about 10% to almost 5x.  Other improvements:
- LastSample: 7.5x
- min_over_time: 6x

Step(s) | Window(s) | Overlap% | SlidingWindow (Row) | ChunkedWindow
-- | -- | -- | -- | --
150 | 300 | 50% | 1000 QPS | 4698 QPS
300 | 300 | 0% |   | 4880 QPS
100 | 300 | 66% |   |  
60 | 300 | 80% | 939 QPS | 2864 QPS
60 | 480 | 88% |   |  
10 | 300 | 97% | 1169 QPS | 1228 QPS
60 | 60 | 0% | 924 QPS | 3223 QPS
60 | 120 | 50% |   | 3227 QPS

In addition, optimizations which benefit even the `SlidingWindowIterator` are included:
- Moving vars from the body of each class to the constructors of inner iterators in various BinaryVectors, as well as the `FilteredChunkInfoIterator`.... saves about 10% 
- Some improvements to SerializableRangeVector
- A fix for the binarySearch functionality used in `PartitionTimeRangeReader` and new chunked iterators