@velvia 2019-01-09
@vishramachandran NaN / EOTS marker handling has been added for all chunked range functions for DoubleColumns.   Here are the before/after comparisons for QueryInMemoryBenchmark.

Before:
```
[info] QueryInMemoryBenchmark.noOverlapQueries     thrpt   20   4341.321 ± 354.499  ops/s
[info] QueryInMemoryBenchmark.singleThreadedQuery  thrpt   20  18654.377 ± 428.580  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries   thrpt   20   3701.860 ± 380.050  ops/s
```

After:
```
[info] QueryInMemoryBenchmark.noOverlapQueries     thrpt   20   4722.348 ± 426.260  ops/s
[info] QueryInMemoryBenchmark.singleThreadedQuery  thrpt   20  19568.348 ± 410.880  ops/s
[info] QueryInMemoryBenchmark.someOverlapQueries   thrpt   20   4278.261 ± 458.860  ops/s
```

No negative effects.  :)
BTW the old sliding iterator I'm not sure handles all NaN / EOTS use cases correctly.  I could not find any test cases for it, I added one.